fix: when consolidating utxos for a msig wallet we consolidate to a primary receive address instead of change so that it shows up in transaction history
fix: many UI updates, sign button is no labelled process as users can add psbt's which are already complete along with raw transactions, or incomplete psbt's. All of which need to processed in order to see what needs to be done to them. New logo, new images.

fix: when exporting files we export to file system not share sheet, this improves UX on macOS

fix: major refactor and hugely improved UX to the "wallet details" view