refactor: major refactor for how FN works with JM
Previously FN would import a duplicate of the jm wallet so that users would get the full FN feature set in a jm wallet including the ability to backup and recover. However it is bad practice to duplicate wallets that watch similiar addresses, it can lead to weird behavior like transactions not showing up in your history. FN now creates a shadow wallet instead and utilizes the join market rpc wallet as per the join market config exclusively. All spend/receive functionality in a join market wallet will occur via the join market api *not* bitcoin core directly. This means you will not see the transaction verifier as jm does not utilize psbts yet, amongst other unimportnant things. This is an improvement as it simplifies the app, makes it lighter and less likely to produce a bug while maintaning functionality.

Proper recovery of your jm wallet should be done on JM! You should only use FN as a medium to an existing jm wallet, do not rely on fn to recover jm wallets for you. The reasons for this are: jm fidelity addresses are only spendable by jm and the jm api does not allow for proper wallet recovery, only the ability to unlock an existing wallet or create a new one.

Users will always be prompted to choose a mixdepth from 0 to 4 to either spend from or receive to when using a jm wallet, if no jm node is active it will fail.

When viewing the jm utxos we use the jm api to fetch those utxos for that specific jm wallet, not the rpc wallet.

We no longer keep utxo labels, we use address labeling by bitcoin core and our own labels/memos for transaction ids, you can use the tx memo/label for utxos.

The gap limit is now 100, we only return 100 transactions and 100 utxos at most.

Esplora is gone, fetching historic fiat is removed, icloud is gone.

Only show onchain balance if no offchain node is active.

getaddressinfo is now called when a recipient address is derived from a wallet, the solvable and path values are visible if sending to the same wallet.